Add showHelpOnFail(false) back; Use middleware function to set handled
showHelpOnFail(false) is needed otherwise it gets quite annoying if a
single command is wrong. Using a middleware function lets us easily set
the handled variable to true if a command was matched.

This means we will only show the entire help text if an unknown command
is entered, matching the old behaviour.